Add of 'graphql fragment' like using const
