+++ src/queries/payments.ts
@@ -1,19 +1,11 @@
 import { processRequest } from '../utils/graphql';
+import { paymentsFields } from './fieldTypes';
 
 export const getPaymentsByService = (serviceId: string): Promise<any> => {
   const query = `
     {
       payments(where: { service: "${serviceId}" }, orderBy: id, orderDirection: asc) {
-        id
-        amount
-        rateToken {
-          address
-          decimals
-          name
-          symbol
-        }
-        paymentType
-        transactionHash
+        ${paymentsFields}
       }
     }
     `;
@@ -24,20 +16,7 @@ export const getPaymentsForUser = (userId: string): Promise<any> => {
   const query = `
     {
       payments(where: {service_: {seller: "${userId}"} }){
-        id, 
-        rateToken {
-          address
-          decimals
-          name
-          symbol
-        }
-        amount
-        transactionHash
-        paymentType
-        service {
-          id, 
-          cid
-        }
+        ${paymentsFields}
       }
     }
