+++ src/queries/proposals.ts
@@ -1,44 +1,11 @@
 import { processRequest } from '../utils/graphql';
+import { proposalsFields } from './fieldTypes';
 
 export const getAllProposalsbyServiceId = (id: string): Promise<any> => {
   const query = `
     {
       proposals(where: {service_: {id: "${id}"}}) {
-        service {
-          id,
-          cid
-          buyer {
-            id
-          }
-        }
-        cid
-        id
-        status
-        rateToken {
-          address
-          decimals
-          name
-          symbol
-        }
-        rateAmount
-        createdAt
-        updatedAt
-        seller {
-          id
-          handle
-          address
-          cid
-          withPoh
-          rating
-          numReviews
-        }
-        description {
-          title
-          id
-          about
-          expectedHours
-          startDate
-        }
+        ${proposalsFields}
       }
     }
     `;
@@ -49,37 +16,7 @@ export const getAllProposalsByUser = (id: string): Promise<any> => {
   const query = `
       {
         proposals(where: {seller: "${id}", status: "Pending"}) {
-          id
-          rateAmount
-          rateToken {
-            address
-            decimals
-            name
-            symbol
-          }
-          status
-          cid
-          createdAt
-          seller {
-            id
-            handle
-          }
-          service {
-            id
-            cid
-            createdAt
-            buyer {
-              id
-              handle
-            }
-          }
-          description {
-            title
-            id
-            about
-            expectedHours
-            startDate
-          }
+          ${proposalsFields}
         }
       }
     `;
