+++ src/queries/reviews.ts
@@ -1,24 +1,11 @@
 import { processRequest } from '../utils/graphql';
+import { reviewsFields } from './fieldTypes';
 
 export const getReviewsByService = (serviceId: string): Promise<any> => {
   const query = `
     {
       reviews(where: { service: "${serviceId}" }, orderBy: id, orderDirection: desc) {
-        id
-        rating
-        createdAt
-        service {
-          id
-          status
-        }
-        to {
-          id
-          handle
-        }
-        description{
-          id
-          content
-        }
+        ${reviewsFields}
       }
     }
     `;
